requirements.txt
openai-whisper
sounddevice  
pynput
scipy
numpy
certifi